Move users endpoint into contact controller
Make consistent with projects resource implementation, move contact serialization responsibilities into own controller.
