+++ server/src/contact/contact.module.ts
@@ -1,6 +1,7 @@
 import { Module } from '@nestjs/common';
 import { CrmModule } from '../crm/crm.module';
 import { ContactService } from './contact.service';
+import { ContactController } from './contact.controller';
 
 @Module({
   imports: [
@@ -8,5 +9,6 @@ import { ContactService } from './contact.service';
   ],
   providers: [ContactService],
   exports: [ContactService],
+  controllers: [ContactController],
 })
 export class ContactModule {}
